Update email sending service file paths
