+++ DotnetFoundation/DotnetFoundation.Domain/Enums/ChannelType.cs
@@ -1,0 +1,8 @@
+﻿namespace DotnetFoundation.Domain.Enums;
+
+public enum ChannelType
+{
+    SMTP = 1,
+    MailGun = 2,
+    WhatsApp = 3
+}
