+++ DotnetFoundation/DotnetFoundation.Domain/Enums/EmailEvents.cs
@@ -1,0 +1,6 @@
+﻿namespace DotnetFoundation.Domain.Enums;
+
+public enum EmailEvents
+{
+    ForgetPassword
+}
