+++ DotnetFoundation/DotnetFoundation.Services/DependencyInjection.cs
@@ -1,6 +1,5 @@
 ﻿using DotnetFoundation.Application.Interfaces.Services;
 using DotnetFoundation.Services.Services.Authentication;
-using DotnetFoundation.Services.Services.EmailService;
 using DotnetFoundation.Services.Services.UserService;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -12,7 +11,6 @@ public static class DependencyInjection
         // Configure service scope for services / BLLs
         services.AddScoped<IAuthenticationService, AuthenticationService>();
         services.AddScoped<IUserService, UserService>();
-        services.AddScoped<IEmailService, EmailService>();
 
         return services;
     }
